requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.2.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.4
itsdangerous==2.2.0
cryptography==43.0.3
python-dotenv==1.0.1
WTForms==3.1.2
Flask-Mail==0.9.1
